Fixed explosion animations playing for some enemy pawns
Adding a SpaceScript directly to a DamageList causes it to have an explosion animation attached to it by the game. Using AddScript on a skill effect doesn't do that.
Since it always uses the pawn's explosion animation, the game probably uses GetExplosion() function, so maybe we could temporarily set the pawn's Explosion field to an empty string?