Fix Pawn issue
Remove code in modApiExt that attempts to update the global Pawn when using weapons.

As a result of this, the global Pawn will only be set by the vanilla game, and will not be usable in queued attacks.

This fixes weapons that uses `Pawn` and has movement as part of their attack from erroring in test mech scenario