Add support for TipImageCallbacks + Refactor
### TipImage Callbacks
This feature allows the optional definiton of a `TipImageCallbacks` table for a weapon, which associates specific elements in the weapon's `TipImage` table to callback functions in `TipImageCallbacks`.

Example:
```lua
TipImageCallbacks = {
	ConveyorUp = function(loc)
		Board:AddAnimation(loc, "Conveyor_0")
	end,
	ConveyorDown = function(loc)
		Board:AddAnimation(loc, "Conveyor_2")
	end,
}
```
This example shows a `TipImageCallbacks` table with two custom callbacks that can be referenced by the weapon's `TipImage`; either by the same name or the name with a numbered suffix.

&nbsp;

Example Weapon:
```lua
	TipImage = {
		Unit = Point(2,2),
		Enemy = Point(2,4),
		Target = Point(2,4),
		ConveyorUp = Point(1,3),
		ConveyorUp2 = Point(1,4),
	},
	TipImageCallbacks = {
		ConveyorUp = function(loc)
			Board:AddAnimation(loc, "Conveyor_0")
		end,
	},
```
In this example, the weapon "MySkill" has a `TipImage` table with a `ConveyorUp` at `Point(1,3)`. When the tipimage is displayed in-game, the `TipImageCallbacks.ConveyorUp` function will be called for the `TipImage.ConveyorUp` element, causing an animation "Conveyor_0" to be created at `Point(1,3)` on the game board.

It also has a second `ConveyorUp` entry, suffixed by `2`. The callback TipImageCallbacks.ConveyorUp will be called for this entry as well.