Fetch board data on the start of a mission
This is a functional change, but I believe one that makes more sense. It could potentially cause mods that uses the library to have to update, in the event that they relied on board events on board init.

This change will for example make it so you will no longer see BuildingCreated or ItemCreated events on init for buildings that are already on the board. Same goes for all other events that are dispatched due to changes of the board.

If mods relied on this to for example find all buildings on the board, they would have to update to fetch all the buildings on the board on mission start instead.